+++ tests/integration/plugins/test_maven_use.py
@@ -66,7 +66,7 @@ def test_maven_use_plugin_self_contained(new_dir, partitions, monkeypatch, caplo
 
     lf = LifecycleManager(
         parts,
-        application_name="test_go",
+        application_name="test_maven_use_self_contained",
         cache_dir=new_dir,
         work_dir=work_dir,
         partitions=partitions,
@@ -113,6 +113,72 @@ def test_maven_use_plugin_self_contained(new_dir, partitions, monkeypatch, caplo
         f"Setting version of 'org.apache.maven.plugins.maven-shade-plugin' to '{shaded_version}'"
         in log
     )
+
+
+def test_maven_use_with_modules(
+    new_dir: Path,
+    partitions: list[str],
+    monkeypatch: pytest.MonkeyPatch,
+    caplog: pytest.LogCaptureFixture,
+) -> None:
+    caplog.set_level(logging.DEBUG)
+    project_dir = Path(new_dir) / "project"
+    shutil.copytree(SOURCE_DIR / "multi-module", project_dir)
+    monkeypatch.chdir(project_dir)
+    work_dir = Path(new_dir)
+    parts_yaml = (project_dir / "parts.yaml").read_text()
+    parts = yaml.safe_load(parts_yaml)
+
+    lf = LifecycleManager(
+        parts,
+        application_name="test_maven_use_with_modules",
+        cache_dir=new_dir,
+        work_dir=work_dir,
+        partitions=partitions,
+    )
+    actions = lf.plan(Step.PRIME)
+
+    with lf.action_executor() as ctx:
+        ctx.execute(actions)
+
+    jar_dir = lf.project_info.prime_dir / "jar"
+    jars = {i.name for i in jar_dir.iterdir()}
+    # Only the two jars produced by the main part
+    # (the shade plugin creates the "original-*" one)
+    assert jars == {"hello-world-0.1.0.jar", "original-hello-world-0.1.0.jar"}
+
+    log = caplog.text
+    assert "Setting version of 'org.starcraft.subsubmod' to '1.0.0'" in log
+
+    assert (
+        "Discovered poms for part 'java-dep-top': [pom.xml, java-dep-submod/pom.xml, java-dep-submod/java-dep-subsubmod/pom.xml]"
+        in log
+    )
+    assert "Discovered poms for part 'java-main-part': [pom.xml]" in log
+
+    maven_repo = lf.project_info.dirs.backstage_dir / "maven-use"
+    bs_dirs: set[str] = set()
+    bs_jars: set[str] = set()
+    # python>=3.12 needed for Path.walk()
+    for node in maven_repo.rglob("*"):
+        rel_node = str(node.relative_to(maven_repo))
+        if node.is_dir():
+            bs_dirs.add(rel_node)
+        elif node.suffix == ".jar":
+            bs_jars.add(rel_node)
+
+    # What we built, and nothing more, should be in the backstage
+    assert bs_dirs == {
+        "org",
+        "org/starcraft",
+        "org/starcraft/top",
+        "org/starcraft/top/1.0.0",
+        "org/starcraft/submod",
+        "org/starcraft/submod/1.0.0",
+        "org/starcraft/subsubmod",
+        "org/starcraft/subsubmod/1.0.0",
+    }
+    assert bs_jars == {"org/starcraft/subsubmod/1.0.0/subsubmod-1.0.0.jar"}
 
 
 @pytest.fixture